+++ src/modules/commands/krillAbout.py
@@ -4,6 +4,7 @@ from datetime import datetime
 import time as PyTime
 from modules.commands.krillVersion import getCurVersion
 from modules.backend.betterLogs.betterLogs import *
+from modules.backend.getChangelog import get_changelog
 from inspect import currentframe, getframeinfo
 readme = ''
 privacyPolicy = ''
@@ -27,24 +28,8 @@ except Exception as e: print('COULD NOT GET FILENAME "' + str(e) + '"')
 PyTime.sleep(1)
 
 def make_changelog():
-    changelog = versionString + '''
-
-### Changed
-
-- Made it to where the PyNaCl error doesn't come before the official start of the log, if it occurs (due to user not having it installed).
-- Made it to where there isnt a weird space between `<!-- Created by Krill You Bot v(Version)-->` and `<!-- Log Generator: "Better Logs V2" | Better Logs by Char @annyconducter on Discord | https://github.com/CharGoldenYT/betterLogs -->` in the log file.
-- Added a little check to make sure the channel i use doesn't receive update announcements.
-- Made the update command DM the user if they try to broadcast `?krill version`.
-- Made the version command suppress embeds on broadcast, similarly to how it works normally when not broadcasted.
-- Further fixed me forgetting to add required function arguments.
-- Made Better Logs handle only having one argument
-- Making bigger, but slower produced updates here on out. So i dont rush it out the door a buggy mess lmao.
-- Fixed the fact that the changelog is actually slightly incorrect.
-- Made it not print so much unneccassary data.
-- Removed old code from `src/modules/backend/exitTasks.py`
-- Removed printing returns of grabbed github files (e.g. the Readme.)
-- Fixed the title not being able to be renamed.
-- Fixed the colors of the readme check.'''
+    r'''Redirect function cause im too lazy :3'''
+    changelog = get_changelog()
     return changelog
 
 # author: Username of who ran the command
@@ -62,8 +47,7 @@ def get_readme(showGetReturns:bool):
     url = ''
     try:url = str(urllib.urlopen('https://raw.githubusercontent.com/gameygu-0213/KrillYouBot/main/readmes/discord_readme.md').read().decode('utf-8'))
     except urllib.HTTPError as e: 
-        frameinfo = getframeinfo(currentframe()); log_err(filname, '[' + str(frameinfo.filename) + '] [' + str(frameinfo.lineno) + ']shit the readme url handler died lmao: ' + str(e))
-        frameinfo = getframeinfo(currentframe()); print('[' + str(frameinfo.filename) + '] [' + str(frameinfo.lineno) + '] shit the readme url handler died lmao: ' + str(e)); url = '-# URL Handler died lmao. '
+        frameinfo = getframeinfo(currentframe()); log_err(filname, '[' + str(frameinfo.filename) + '] [' + str(frameinfo.lineno) + '] shit the readme url handler died lmao: ' + str(e)); url = '-# URL Handler died lmao. '
     readme = 'This was generated with the [GitHub "discord_readme"](https://github.com/gameygu-0213/KrillYouBot/blob/main/readmes/discord_readme.md):\n\n' + '# Krill You Bot' + versionString + ' ' + url
 
     #if showGetReturns:print('get_readme returned: ' + readme)
@@ -73,8 +57,7 @@ def get_privacy_policy(showGetReturns:bool):
     url = ''
     try:url = str(urllib.urlopen('https://raw.githubusercontent.com/CharGoldenYT/KrillYouBot/main/readmes/Privacy%20Policy.md').read().decode('utf-8'))
     except urllib.HTTPError as e: 
-        frameinfo = getframeinfo(currentframe()); log_err(filname, '[' + str(frameinfo.filename) + '] [' + str(frameinfo.lineno) + ']shit the readme url handler died lmao: ' + str(e))
-        frameinfo = getframeinfo(currentframe()); print('[' + str(frameinfo.filename) + '] [' + str(frameinfo.lineno) + '] shit the readme url handler died lmao: ' + str(e)); url = '-# URL Handler died lmao. '
+        frameinfo = getframeinfo(currentframe()); log_err(filname, '[' + str(frameinfo.filename) + '] [' + str(frameinfo.lineno) + '] shit the readme url handler died lmao: ' + str(e)); url = '-# URL Handler died lmao. '
     privacyPolicy = 'This was generated with the [GitHub "Privacy Policy"](https://github.com/gameygu-0213/KrillYouBot/blob/main/readmes/Privacy%20Policy.md):\n\n' + url
 
     #if showGetReturns:print('get_privacy_policy returned: ' + privacyPolicy)
@@ -84,8 +67,7 @@ def get_tos(showGetReturns:bool):
     url = ''
     try:url = str(urllib.urlopen('https://raw.githubusercontent.com/gameygu-0213/KrillYouBot/main/readmes/tos.md').read().decode('utf-8'))
     except urllib.HTTPError as e: 
-        frameinfo = getframeinfo(currentframe()); log_err(filname, '[' + str(frameinfo.filename) + '] [' + str(frameinfo.lineno) + ']shit the readme url handler died lmao: ' + str(e))
-        frameinfo = getframeinfo(currentframe()); print('[' + str(frameinfo.filename) + '] [' + str(frameinfo.lineno) + '] shit the readme url handler died lmao: ' + str(e)); url = '-# URL Handler died lmao. '
+        frameinfo = getframeinfo(currentframe()); log_err(filname, '[' + str(frameinfo.filename) + '] [' + str(frameinfo.lineno) + '] shit the readme url handler died lmao: ' + str(e)); url = '-# URL Handler died lmao. '
     tos = 'This was generated with the [GitHub "tos"](https://github.com/gameygu-0213/KrillYouBot/blob/main/readmes/tos.md):\n\n' + url
 
     #if showGetReturns:print('get_readme returned: ' + tos)
@@ -96,7 +78,6 @@ def get_gitVer():
     versionToReturn = None
     try:url = str(urllib.urlopen('https://raw.githubusercontent.com/gameygu-0213/KrillYouBot/main/gitVer.txt').read().decode('utf-8'))
     except urllib.HTTPError as e: 
-        frameinfo = getframeinfo(currentframe()); log_err(filname, '[' + str(frameinfo.filename) + '] [' + str(frameinfo.lineno) + ']shit the readme url handler died lmao: ' + str(e))
-        frameinfo = getframeinfo(currentframe()); print('[' + str(frameinfo.filename) + '] [' + str(frameinfo.lineno) + '] shit the readme url handler died lmao: ' + str(e))
+        frameinfo = getframeinfo(currentframe()); log_err(filname, '[' + str(frameinfo.filename) + '] [' + str(frameinfo.lineno) + '] shit the readme url handler died lmao: ' + str(e))
     if not url == None:versionToReturn = url
     return versionToReturn