+++ readmes/changelog.md
@@ -1,5 +1,3 @@
-# TODO: MAKE THIS ABLE TO BE READ BY KRILL YOU BOT SO I DONT HAVE TO MANUALLY RETYPE THIS EVERYTIME.
-
 # Changelog
 
 All notable changes to this project will be documented in this file.
@@ -27,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed printing returns of grabbed github files (e.g. the Readme.)
 - Fixed the title not being able to be renamed.
 - Fixed the colors of the readme check.
+- Removed duplicate prints in `src/modules/commands/krillAbout.py`
+- Made it possible to retrieve the current changelog from gihub.
 
 ## [3.2h-2] 10/18/24 - 2:29 AM